cursor_mcp_requirements.txt
# Cursor Background Agent MCP Server Requirements

# Core MCP dependencies
mcp>=1.0.0

# HTTP client for Cursor API
aiohttp>=3.8.0
aiofiles>=23.0.0

# JSON schema validation
jsonschema>=4.17.0

# Environment variable management  
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Optional: Enhanced HTTP client features
httpx>=0.24.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.0.0
